// src/payment-methods/dto/update-payment-method.dto.ts
export class UpdatePaymentMethodDto {
  description?: string;
  category?: string;
  currency?: string;
  generatesChange?: boolean;
  invoiceCash?: boolean;
  invoiceCredit?: boolean;
  fuelOutflow?: boolean;
  loyalty?: boolean;
  requiresReference?: boolean;
  image?: string;
  active?: boolean;
  accountId?: string;
  commissionPct?: number;
  storeCodes?: string[];
}